+++ history.md
@@ -3,6 +3,7 @@
 * feat: webpack 类型项目支持配置中设置  `devServer` 配置项
 * feat: 让 `devServer.historyApiCallback` 属性能在 proxy 模式下使用
 * feat: 优化 error log
+* feat: webpack 类项目 `entry/*.html`  `entry/*.pug` 支持 变量传入
 * fix: 修复 webpack 类项目 sugar 执行不符合预期的问题
 
 ## 3.9.0-beta2 (2020-02-24)